requirements.txt
numpy>=1.24.0
wandb>=0.15.0
deepspeed>=0.10.0
tqdm>=4.65.0
pyyaml>=6.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
lightning>=2.0.0
tensorboard>=2.11.0

# torch>=2.0.0
# torchvision>=0.15.0
# flash-attn>=2.0.0
# h5py>=3.8.0